+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## 0.2.0 - UNRELEASED
 
+Requires `libfranka` >= 0.2.0
+
   * Added missing run-time dependencies to `franka_description` and `franka_control`
-  * Added `m_ee`, `F_x_Cee`, `I_ee`, `m_total`, `F_x_Ctotal` and `I_total`
-    to the robot state
-  * Updated and improved examples
+  * Added `tau_J_d`, `m_ee`, `F_x_Cee`, `I_ee`, `m_total`, `F_x_Ctotal`, `I_total`,
+    `theta` and `dtheta` to `franka_msgs/FrankaState`
+  * Added new errors to `franka_msgs/Errors`
+  * Updated and improved examples in `franka_example_controllers`
   * Fixed includes for Eigen3 in `franka_example_controllers`
   * Fixed gripper state publishing in `franka_gripper_node`